settlement create test

+++ contracts/src/components.cairo
@@ -1,7 +1,7 @@
 #[derive(Component, Copy, Drop, Serde, SerdeLen)]
 struct Position {
     #[key]
-    entity_id: u32,
+    entity_id: u128,
     #[key]
     game_id: u32,
     x: u32,
@@ -11,7 +11,7 @@ struct Position {
 #[derive(Component, Copy, Drop, Serde, SerdeLen)]
 struct Lifes {
     #[key]
-    entity_id: u32,
+    entity_id: u128,
     #[key]
     game_id: u32,
     count: u32
@@ -20,7 +20,7 @@ struct Lifes {
 #[derive(Component, Copy, Drop, Serde, SerdeLen)]
 struct Defence {
     #[key]
-    entity_id: u32,
+    entity_id: u128,
     #[key]
     game_id: u32,
     plague: u32
@@ -29,7 +29,7 @@ struct Defence {
 #[derive(Component, Copy, Drop, Serde, SerdeLen)]
 struct Name {
     #[key]
-    entity_id: u32,
+    entity_id: u128,
     #[key]
     game_id: u32,
     value: felt252
@@ -38,7 +38,7 @@ struct Name {
 #[derive(Component, Copy, Drop, Serde, SerdeLen)]
 struct Prosperity {
     #[key]
-    entity_id: u32,
+    entity_id: u128,
     #[key]
     game_id: u32,
     value: felt252
@@ -47,7 +47,7 @@ struct Prosperity {
 #[derive(Component, Copy, Drop, Serde, SerdeLen)]
 struct Balance {
     #[key]
-    entity_id: u32,
+    entity_id: u128,
     #[key]
     game_id: u32,
     value: felt252
@@ -57,7 +57,7 @@ struct Balance {
 #[derive(Component, Copy, Drop, Serde, SerdeLen)]
 struct WorldEvent {
     #[key]
-    entity_id: u32,
+    entity_id: u128,
     #[key]
     game_id: u32,
     radius: u32,
@@ -81,14 +81,14 @@ struct Game {
 #[derive(Component, Copy, Drop, Serde, SerdeLen)]
 struct GameTracker {
     #[key]
-    entity_id: u32, // FIXED
+    entity_id: u128, // FIXED
     count: u32
 }
 
 #[derive(Component, Copy, Drop, Serde, SerdeLen)]
 struct Ownership {
     #[key]
-    entity_id: u32, // FIXED
+    entity_id: u128, // FIXED
     #[key]
     game_id: u32, // increment
     address: felt252
